+++ src/reader/reader.rs
@@ -75,8 +75,7 @@ use crate::fallback::fish_wcwidth;
 use crate::fd_readable_set::poll_fd_readable;
 use crate::fds::{AutoCloseFd, make_fd_blocking, wopen_cloexec};
 use crate::flog::{FLOG, FLOGF};
-use crate::future_feature_flags;
-use crate::future_feature_flags::FeatureFlag;
+use crate::future_feature_flags::{self, FeatureFlag};
 use crate::global_safety::RelaxedAtomicBool;
 use crate::highlight::{
     HighlightRole, HighlightSpec, autosuggest_validate_from_history, highlight_shell,